Introduced telnet session detection..
This is a try to deal with unexpected "..11.." or "..22.." in V4log1 file..
The created V4.exe not ran at the beginning on SGLmachine1 (10.36.231.115).. It was necessary to reinstall the Cygwin
with the same version as on the machine where V4.exe was built. There was weird errors.. look for
tag_issue CanNotLocatePAR_pmInINC,CanNotLoadDLLForModule in D:\Profiles\Administrator\MyDocuments\stuff\tmp2.txt
V4.exe was obtained like this: "pp -o ./V4.exe -n -x ./V4.pl --xargs=dxt2".